refactor: remove dead code
