+++ src/nitpick/generic.py
@@ -8,53 +8,10 @@ import re
 from pathlib import Path
 from typing import Iterable, List, Optional, Tuple
 
-from nitpick.constants import DOUBLE_QUOTE, PROJECT_NAME, SEPARATOR_QUOTED_SPLIT, SINGLE_QUOTE
+from nitpick.constants import PROJECT_NAME
 from nitpick.typedefs import PathOrStr
 
 URL_RE = re.compile(r"[a-z]+://[\$\w]\w*")
-
-
-def get_subclasses(cls):
-    """Recursively get subclasses of a parent class."""
-    subclasses = []
-    for subclass in cls.__subclasses__():
-        subclasses.append(subclass)
-        subclasses += get_subclasses(subclass)
-    return subclasses
-
-
-def quoted_split(string_: str, separator=".") -> List[str]:
-    """Split a string by a separator, but considering quoted parts (single or double quotes).
-
-    >>> quoted_split("my.key.without.quotes")
-    ['my', 'key', 'without', 'quotes']
-    >>> quoted_split('"double.quoted.string"')
-    ['double.quoted.string']
-    >>> quoted_split('"double.quoted.string".and.after')
-    ['double.quoted.string', 'and', 'after']
-    >>> quoted_split('something.before."double.quoted.string"')
-    ['something', 'before', 'double.quoted.string']
-    >>> quoted_split("'single.quoted.string'")
-    ['single.quoted.string']
-    >>> quoted_split("'single.quoted.string'.and.after")
-    ['single.quoted.string', 'and', 'after']
-    >>> quoted_split("something.before.'single.quoted.string'")
-    ['something', 'before', 'single.quoted.string']
-    """
-    if DOUBLE_QUOTE not in string_ and SINGLE_QUOTE not in string_:
-        return string_.split(separator)
-
-    quoted_regex = re.compile(
-        f"([{SINGLE_QUOTE}{DOUBLE_QUOTE}][^{SINGLE_QUOTE}{DOUBLE_QUOTE}]+[{SINGLE_QUOTE}{DOUBLE_QUOTE}])"
-    )
-
-    def remove_quotes(match):
-        return match.group(0).strip("".join([SINGLE_QUOTE, DOUBLE_QUOTE])).replace(separator, SEPARATOR_QUOTED_SPLIT)
-
-    return [
-        part.replace(SEPARATOR_QUOTED_SPLIT, separator)
-        for part in quoted_regex.sub(remove_quotes, string_).split(separator)
-    ]
 
 
 def version_to_tuple(version: str = None) -> Tuple[int, ...]:
