+++ src/nitpick/plugins/yaml.py
@@ -1,11 +1,8 @@
 """YAML files."""
-from collections import OrderedDict
 from itertools import chain
-from typing import Iterator, Optional, Tuple, Type, Union, cast
+from typing import Iterator, Optional, Type, cast
 
-import attr
-
-from nitpick.blender import BaseDoc, YamlDoc, search_json, traverse_yaml_tree
+from nitpick.blender import BaseDoc, YamlDoc, traverse_yaml_tree
 from nitpick.constants import PRE_COMMIT_CONFIG_YAML
 from nitpick.exceptions import Deprecation
 from nitpick.plugins import hookimpl
@@ -20,52 +17,6 @@ KEY_YAML = "yaml"
 KEY_HOOKS = "hooks"
 KEY_REPO = "repo"
 KEY_ID = "id"
-
-
-@attr.s()
-class PreCommitHook:
-    """A pre-commit hook.
-
-    .. note::
-
-        This is a class from the deprecated ``nitpick.plugins.pre_commit.PreCommitPlugin``.
-        It's not being used at the moment, but it was kept here because it's harmless.
-    """
-
-    repo = attr.ib(type=str)
-    hook_id = attr.ib(type=str)
-    yaml = attr.ib(type=YamlDoc)
-
-    @property
-    def unique_key(self) -> str:
-        """Unique key of this hook, to be used in a dict."""
-        return f"{self.repo}_{self.hook_id}"
-
-    @property
-    def key_value_pair(self) -> Tuple[str, "PreCommitHook"]:
-        """Key/value pair to be used as a dict item."""
-        return self.unique_key, self
-
-    @property
-    def single_hook(self) -> JsonDict:
-        """Return only a single hook instead of a list."""
-        return self.yaml.as_list[0]
-
-    @classmethod
-    def get_all_hooks_from(cls, str_or_yaml: Union[str, YamlObject]):
-        """Get all hooks from a YAML string. Split the string in hooks and copy the repo info for each."""
-        yaml = YamlDoc(string=str_or_yaml).as_list if isinstance(str_or_yaml, str) else str_or_yaml
-        hooks = []
-        for repo in yaml:
-            for index, hook in enumerate(repo.get(KEY_HOOKS, [])):
-                repo_data_only = repo.copy()
-                repo_data_only.pop(KEY_HOOKS)
-                hook_data_only = search_json(repo, f"{KEY_HOOKS}[{index}]", {})
-                repo_data_only.update({KEY_HOOKS: [hook_data_only]})
-                hooks.append(
-                    PreCommitHook(repo.get(KEY_REPO), hook[KEY_ID], YamlDoc(obj=[repo_data_only])).key_value_pair  # type: ignore[arg-type]
-                )
-        return OrderedDict(hooks)
 
 
 class YamlPlugin(NitpickPlugin):
